refactor(backup-client): move the backup client to the remote client
Previously it was part of the local client which was wrong, given that the backup endpoints are exposed on the remote webserver
